requirements.txt
you-get
openai-whisper
spleeter==2.3.0
moviepy
ffmpeg-python
# audio-slicer
appdirs==1.4.4
audioread==3.0.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.1
decorator==5.1.1
idna==3.3
importlib-metadata==4.12.0
joblib==1.1.0
numba
packaging==21.3
pooch==1.6.0
pycparser==2.21
pyparsing==3.0.9
requests==2.28.1
resampy==0.4.0
scikit-learn==1.1.2
scipy==1.9.1
SoundFile==0.10.3.post1
threadpoolctl==3.1.0
urllib3==1.26.12
zipp==3.8.1